+++ lib/src/ui/pages/cupertino/activity_indicator.g.dart
@@ -1,0 +1,25 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'activity_indicator.dart';
+
+// **************************************************************************
+// RiverpodGenerator
+// **************************************************************************
+
+String _$configHash() => r'791b148465e87ac927ad36fe39b5453579b5b5b2';
+
+/// See also [Config].
+@ProviderFor(Config)
+final configProvider =
+    AutoDisposeNotifierProvider<Config, ActivityIndocatorConfig>.internal(
+  Config.new,
+  name: r'configProvider',
+  debugGetCreateSourceHash:
+      const bool.fromEnvironment('dart.vm.product') ? null : _$configHash,
+  dependencies: null,
+  allTransitiveDependencies: null,
+);
+
+typedef _$Config = AutoDisposeNotifier<ActivityIndocatorConfig>;
+// ignore_for_file: type=lint
+// ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member