+++ lib/src/ui/pages/cupertino/alert_dialog.g.dart
@@ -1,0 +1,25 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'alert_dialog.dart';
+
+// **************************************************************************
+// RiverpodGenerator
+// **************************************************************************
+
+String _$configHash() => r'aa0aa6326a2787bcdfb615a684826c4979b89d40';
+
+/// See also [Config].
+@ProviderFor(Config)
+final configProvider =
+    AutoDisposeNotifierProvider<Config, AlertDialogConfig>.internal(
+  Config.new,
+  name: r'configProvider',
+  debugGetCreateSourceHash:
+      const bool.fromEnvironment('dart.vm.product') ? null : _$configHash,
+  dependencies: null,
+  allTransitiveDependencies: null,
+);
+
+typedef _$Config = AutoDisposeNotifier<AlertDialogConfig>;
+// ignore_for_file: type=lint
+// ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member