+++ lib/src/ui/pages/cupertino/widgets.dart
@@ -1,11 +1,19 @@
 import 'package:dreambook/src/ui/pages/shared/shared_code_view.dart';
 
+import 'action_sheet.dart';
+import 'activity_indicator.dart';
+import 'alert_dialog.dart';
+import 'button.dart';
 import 'list_section.dart';
 import 'list_tile.dart';
 import 'slider.dart';
 import 'switch.dart';
 
 final cupertinoItems = <CodeItem>[
+  actionSheetItem,
+  activityIndicatorItem,
+  alertDialogItem,
+  buttonItem,
   switchItem,
   sliderItem,
   listTileItem,